requirements.txt
click==8.1.3
requests==2.28.2
pyjwt==2.6.0
cryptography==39.0.0
